helpers refactoring
